+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.5.1-beta.x
 
-- Add `isBigInt` to allow `BigInt` checks. Also add support in `bnToBn`
+- Add `isBigInt(value)` to allow `BigInt` checks. Also add support in `bnToBn`, and anything that uses this conversion for numbers
+- Add `isChildClass(Parent, Child?)` to check if a class extends the parent
 
 # 2.4.1 Feb 18, 2020