+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# 2.5.1-beta.x
+# 2.5.1 Feb 22, 2020
 
-- Add `isBigInt(value)` to allow `BigInt` checks. Also add support in `bnToBn`, and anything that uses this conversion for numbers
-- Add `isChildClass(Parent, Child?)` to check if a class extends the parent
+- Add `isBigInt(value)` to allow `BigInt` checks. Also add support in `bnToBn` as well as number conversion utils
+- Add `isChildClass(Parent, Child?)` to check if a class extends the parent, asserting Child is Parent
 - (internal) Use `#<varname>` insead of `private _<varname>` for private class variables
 
 # 2.4.1 Feb 18, 2020